finalizing version 1

+++ app/src/main/java/com/brainup/woyalladriver/Activities/MainActivity.java
@@ -418,34 +418,35 @@ public class MainActivity extends AppCompatActivity
     /**
      * reload the map to clear previous clients and update location
      */
-    public void reload(){
-        //remove all clients
-        WoyallaDriver.myDatabase.Delete_All(Database.Table_CLIENT);
-        client_available.setVisibility(View.GONE);
-        int user_id = WoyallaDriver.myDatabase.get_Top_ID(Database.Table_USER);
-        int currentStatus = Integer.parseInt(WoyallaDriver.myDatabase.get_Value_At_Top(Database.Table_USER,Database.USER_FIELDS[8]));
-        if(currentStatus>1){
-            if(avaialbilitySwitch.isChecked()){
-                //Change the client status to active
-                ContentValues userStatus = new ContentValues();
-                userStatus.put(Database.USER_FIELDS[8],"1");
-                WoyallaDriver.myDatabase.update(Database.Table_USER,userStatus,user_id);
+    public void reload() {
+        if (mMap != null) {
+            //remove all clients
+            WoyallaDriver.myDatabase.Delete_All(Database.Table_CLIENT);
+            client_available.setVisibility(View.GONE);
+            int user_id = WoyallaDriver.myDatabase.get_Top_ID(Database.Table_USER);
+            int currentStatus = Integer.parseInt(WoyallaDriver.myDatabase.get_Value_At_Top(Database.Table_USER, Database.USER_FIELDS[8]));
+            if (currentStatus > 1) {
+                if (avaialbilitySwitch.isChecked()) {
+                    //Change the client status to active
+                    ContentValues userStatus = new ContentValues();
+                    userStatus.put(Database.USER_FIELDS[8], "1");
+                    WoyallaDriver.myDatabase.update(Database.Table_USER, userStatus, user_id);
+                } else {
+                    //Change the client status to offline
+                    ContentValues userStatus = new ContentValues();
+                    userStatus.put(Database.USER_FIELDS[8], "0");
+                    WoyallaDriver.myDatabase.update(Database.Table_USER, userStatus, user_id);
+                }
             }
-            else{
-                //Change the client status to offline
-                ContentValues userStatus = new ContentValues();
-                userStatus.put(Database.USER_FIELDS[8],"0");
-                WoyallaDriver.myDatabase.update(Database.Table_USER,userStatus,user_id);
-            }
-        }
 
-        mMap.clear();   //clear the client marker from the map
-        /**
-         * First get the location data from the database.
-         * then view it on the map
-         */
-        moveMap(MainActivity.this.getResources().getString(R.string.my_location));
-        Toast.makeText(MainActivity.this,MainActivity.this.getResources().getString(R.string.toast_reload),Toast.LENGTH_LONG).show();
+            mMap.clear();   //clear the client marker from the map
+            /**
+             * First get the location data from the database.
+             * then view it on the map
+             */
+            moveMap(MainActivity.this.getResources().getString(R.string.my_location));
+            Toast.makeText(MainActivity.this, MainActivity.this.getResources().getString(R.string.toast_reload), Toast.LENGTH_LONG).show();
+        }
     }
 
     public void logOut(){
